requirements.txt
anthropic>=0.39.0
yfinance>=0.2.40
fastapi>=0.115.0
uvicorn>=0.32.0
jinja2>=3.1.4
python-dotenv>=1.0.1
apscheduler>=3.10.4
pydantic>=2.9.0
httpx>=0.27.0
pytz>=2024.1
sse-starlette>=2.0.0
